Setting up the new context to instance of createHoc in componentWillReceiveProps
In the code for updateClassInstance method in react-dom.development.js file, first
all the lifecycle methods like componentWillReceiveProps, componentWillUpdate,
shouldComponentUpdate are called and in the end of updateClassInstance,
instance.context is set to new context.. so problem here is that in createHoc's
componentWillReceiveProps createState is called which in turn references this.context.
so this will always reference the old context. Since componentWillReceiveProps is
the first method in update cycle.. setting the context in this instance to newContext.