enhancing inject option to inject only part of theme as prop
By default theme is injected as prop to component. But in the options to injectSheet
we can whitelist props to be injected. However, if we want to inject only some part
of theme as prop there is no way. So adding one more key to inject option to give
the flexibility to inject only part of theme.